backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.17.2",
    "async-signals>=0.3.1",
    "bcrypt==3.2.2",
    "fastapi>=0.128.0",
    "fastapi-mail>=1.6.1",
    "httpx>=0.28.1",
    "passlib[bcrypt]>=1.7.4",
    "psycopg[binary]>=3.3.3",
    "pydantic-settings>=2.12.0",
    "pydantic[email]>=2.12.5",
    "pytest>=9.0.2",
    "pytest-mock>=3.15.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.21",
    "python-slugify>=8.0.4",
    "sqlalchemy>=2.0.45",
    "uvicorn>=0.40.0",
]
